src/main/java/core/basesyntax/Calculator.java
package core.basesyntax;

public class Calculator implements Calculate {

    @Override
     public double calculate(double firstValue, double secondValue, char operation) {
        switch (operation) {
            case '+':
                return firstValue + secondValue;
            case '-':
                return firstValue - secondValue;
            case '*':
                return firstValue * secondValue;
            case '/':
                if (secondValue == 0) {
                    throw new ArithmeticException("Divider cannot be zero");
                }
                return firstValue / secondValue;
            case '^':
                return Math.pow(firstValue, secondValue);
            default:
                throw new RuntimeException("Operation is invalid.");
        }
    }
}
